Create ProcrastinationDetectorSensorListener
